Fairly big changes
Extracting the kernel version string from command output is refactored to a method.
This method uses regular expressions so it should be more portable to other distributions or package managers.

Comparing version strings yourself is a nightmare so I used parse_version from pkg_resources in the standard library instead.