Modifed extracting of kernel version strings
Regex in the extractKernelVersionString () method will now allow letters so that the trailing characters in the version are also picked up.

Old: 4.7.2-200
New: 4.7.2-200.fc24.x86_64

This doesn't affect the comparision and parsing of version strings but signing the modules requires
accessing directories with the full kernel name not just the version.