Started switching working directory to the modules directory
This improves error handling in case the modules directory provided by the user is invalid instead of checking it all in one line
when we try to sign the modules.

The sign-file binary is called by path unchecked because it should be reliable as long as the version string is correct
and kernel-devel (dependancy) is installed.